Fix application start condition

+++ launch-weather.sh
@@ -4,10 +4,19 @@
 path=/var/www/weather
 
 # Kill previous process
+ 
 ps=`pgrep node`
-echo kill node process: ${ps}
-kill -15 ${ps}
 
-sleep 5
-# Launch chrisley's weather app
-nohup node ${path}/bin/www >> ${path}/weather.log &
+if [ -z "${ps}" ]; then
+	echo "chrisley's weather app is not running"
+	echo "starting app..."
+	# Launch chrisley's weather app
+	nohup node ${path}/bin/www >> ${path}/weather.log &
+else
+	echo "chrisley's app is running with PID:" ${ps}
+	echo "restarting app..."
+	kill -15 ${ps}
+	sleep 1
+	# Launch chrisley's weather app
+	nohup node ${path}/bin/www >> ${path}/weather.log &
+fi
